requirements.txt
bleach==3.1.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
pkginfo==1.5.0.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.3.1
pytz==2018.7
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.8.0
rsa==4.0
six==1.14.0
tqdm==4.29.1
twine==1.12.1
urllib3==1.24.1
webencodings==0.5.1
